refactor: extract session metadata from App
